requirements.txt
yahoo_fin==0.8.7
pyserial==3.5
